backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pyyaml==6.0.1
python-multipart==0.0.6
websockets==12.0
httpx==0.26.0
kubernetes==29.0.0
prometheus-client==0.19.0
aiohttp==3.9.1
python-dotenv==1.0.0
